requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
channels==4.0.0
channels-redis==4.1.0
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
msgpack==1.0.5
PyJWT==2.7.0
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.7.0
